+++ src/Application/CruiseRuntime.h
@@ -69,6 +69,7 @@ public:
 
     SelectionDecision CurrentSelection() const;
     ActionDecision CurrentMapAction(const MapActionEnvironment& environment) const;
+    ShipContext CurrentMapShipContext() const;
 
     EffectDispatchResult ActivateMapAction(const MapSessionIdentity& identity, MapActionGesture gesture, const MapActionEnvironment& environment);